pyproject.toml
[tool.poetry]
name = "betheMC"
version = "0.1.0"
description = "A choice-based Pokémon adventure game set in the Kanto region"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "bethemc", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
langchain = "^0.1.0"
langchain-community = "^0.0.10"
openai = "^1.0.0"
pyyaml = "^6.0.1"
qdrant-client = "^1.7.0"
python-dotenv = "^1.0.0"
typing-extensions = "^4.8.0"
requests = "^2.31.0"
sentence-transformers = "^4.1.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
pydantic = {extras = ["email"], version = "^2.11.7"}
pydantic-settings = "^2.10.1"
python-multipart = "^0.0.20"
motor = "^3.3.0"
pymongo = "^4.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
ruff = "^0.1.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 